@@ -1,13 +1,22 @@
 apply plugin: 'com.android.application'
+apply plugin: 'com.google.protobuf'
 
 android {
-    compileSdkVersion 23
-    // buildToolsVersion "25.0.2"
+    sourceSets {
+        main {
+            proto {
+                // In addition to the default 'src/main/proto'
+                srcDir '../protos'
+            }
+        }
+    }
+
+    compileSdkVersion 28
 
     defaultConfig {
         applicationId "edu.cmu.cs.gabrielclient"
-        minSdkVersion 15
-        targetSdkVersion 23
+        minSdkVersion 19
+        targetSdkVersion 28
     }
 
     buildTypes {
@@ -16,11 +25,27 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.txt'
         }
     }
+
+    protobuf {
+        protoc {
+            artifact = 'com.google.protobuf:protoc:3.8.0'
+        }
+        generateProtoTasks {
+            all().each { task ->
+                task.builtins {
+                    java {
+                        option "lite"
+                    }
+                }
+            }
+        }
+    }
 }
 
 dependencies {
-    implementation 'com.android.support:appcompat-v7:23.2.0'
-    implementation 'com.android.support:support-v4:23.2.0'
-    implementation 'com.android.support:support-v13:23.2.0'
-    implementation 'com.android.support:design:23.2.0'
+    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'com.android.support:multidex:1.0.3'
+    implementation 'edu.cmu.cs.gabriel:client:0.1.6'
+    implementation 'edu.cmu.cs.gabriel:protocol:0.1.4'
 }
