@@ -45,36 +45,10 @@ public class Const {
     public static int IMAGE_HEIGHT = 360;
     public static String FOCUS_MODE = Camera.Parameters.FOCUS_MODE_CONTINUOUS_VIDEO;
     public static String FLASH_MODE = null;
-    
-    // port protocol to the server
-    public static final int VIDEO_STREAM_PORT = 9098;
-    public static final int ACC_STREAM_PORT = 9099;
-    public static final int AUDIO_STREAM_PORT = 9100;
-    public static final int RESULT_RECEIVING_PORT = 9111;
-    public static final int CONTROL_PORT = 22222;
 
-    // the app name
-    public static final String APP_NAME = "sandwich";
+    public static final int PORT = 9099;
 
-    // load images (JPEG) from files and pretend they are just captured by the camera
-    public static final File TEST_IMAGE_DIR = new File (ROOT_DIR.getAbsolutePath() +
-            File.separator + "images-" + APP_NAME + File.separator);
-
-    // load audio data from file
-    public static final File TEST_AUDIO_FILE = new File (ROOT_DIR.getAbsolutePath() +
-            File.separator + "audio-" + APP_NAME + ".raw");
-
-    // load acc data from file
-    public static final File TEST_ACC_FILE = new File (ROOT_DIR.getAbsolutePath() +
-            File.separator + "acc-" + APP_NAME + ".txt");
-
-    // load acc to image timing mapping data from file
-    public static final File IMAGE_TIMING_FILE = new File (TEST_IMAGE_DIR +
-            File.separator + "timing.txt");
-
-    // may include background pinging to keep network active
-    public static final boolean BACKGROUND_PING = false;
-    public static final int PING_INTERVAL = 20;
+    public static final String ENGINE_NAME = "instruction";
 
     // audio configurations
     public static final int RECORDER_SAMPLERATE = 16000;
@@ -101,17 +75,6 @@ public class Const {
 
     // token size list
     public static final int[] TOKEN_SIZE_LIST = {1};
-
-    // maximum times to ping (for time synchronization
-    public static final int MAX_PING_TIMES = 20;
-
-    // a small number of images used for compression (bmp files), usually a subset of test images
-    // these files are loaded into memory first so cannot have too many of them!
-    public static final File COMPRESS_IMAGE_DIR = new File (ROOT_DIR.getAbsolutePath() +
-            File.separator + "images-" + APP_NAME + "-compress" + File.separator);
-    // the maximum allowed compress images to load
-    public static final int MAX_COMPRESS_IMAGE = 3;
-    
     // result file
     public static final File EXP_DIR = new File(ROOT_DIR.getAbsolutePath() + File.separator + "exp");
 
