@@ -1,5 +1,5 @@
 # Generated by the protocol buffer compiler.  DO NOT EDIT!
-# source: task.proto
+# source: instruction.proto
 
 import sys
 _b=sys.version_info[0]<3 and (lambda x:x) or (lambda x:x.encode('latin1'))
@@ -16,10 +16,10 @@ _sym_db = _symbol_database.Default()
 
 
 DESCRIPTOR = _descriptor.FileDescriptor(
-  name='task.proto',
+  name='instruction.proto',
   package='gabriel',
   syntax='proto3',
-  serialized_pb=_b('\n\ntask.proto\x12\x07gabriel\"\xf2\x01\n\x04Ikea\x12\"\n\x05state\x18\x01 \x01(\x0e\x32\x13.gabriel.Ikea.State\x12\x1e\n\x16\x66rames_with_one_buckle\x18\x02 \x01(\x05\x12\x1f\n\x17\x66rames_with_two_buckles\x18\x03 \x01(\x05\"\x84\x01\n\x05State\x12\t\n\x05START\x10\x00\x12\x0b\n\x07NOTHING\x10\x01\x12\x08\n\x04\x42\x41SE\x10\x02\x12\x08\n\x04PIPE\x10\x03\x12\t\n\x05SHADE\x10\x04\x12\n\n\x06\x42UCKLE\x10\x05\x12\x10\n\x0c\x42LACK_CIRCLE\x10\x06\x12\x0e\n\nSHADE_BASE\x10\x07\x12\x08\n\x04\x42ULB\x10\x08\x12\x0c\n\x08\x42ULB_TOP\x10\t\"\xa6\x01\n\x08Sandwich\x12&\n\x05state\x18\x01 \x01(\x0e\x32\x17.gabriel.Sandwich.State\"r\n\x05State\x12\t\n\x05START\x10\x00\x12\x0b\n\x07NOTHING\x10\x01\x12\t\n\x05\x42READ\x10\x02\x12\x07\n\x03HAM\x10\x03\x12\x0b\n\x07LETTUCE\x10\x04\x12\x0c\n\x08\x43UCUMBER\x10\x05\x12\x08\n\x04HALF\x10\x06\x12\n\n\x06TOMATO\x10\x07\x12\x0c\n\x08HAMWRONG\x10\x08\"{\n\x0c\x45ngineFields\x12\x1c\n\x14last_update_frame_id\x18\x01 \x01(\x03\x12\x1d\n\x04ikea\x18\x02 \x01(\x0b\x32\r.gabriel.IkeaH\x00\x12%\n\x08sandwich\x18\x03 \x01(\x0b\x32\x11.gabriel.SandwichH\x00\x42\x07\n\x05stateB%\n\x1b\x65\x64u.cmu.cs.gabriel.protocolB\x06Protosb\x06proto3')
+  serialized_pb=_b('\n\x11instruction.proto\x12\x07gabriel\"\xf2\x01\n\x04Ikea\x12\"\n\x05state\x18\x01 \x01(\x0e\x32\x13.gabriel.Ikea.State\x12\x1e\n\x16\x66rames_with_one_buckle\x18\x02 \x01(\x05\x12\x1f\n\x17\x66rames_with_two_buckles\x18\x03 \x01(\x05\"\x84\x01\n\x05State\x12\t\n\x05START\x10\x00\x12\x0b\n\x07NOTHING\x10\x01\x12\x08\n\x04\x42\x41SE\x10\x02\x12\x08\n\x04PIPE\x10\x03\x12\t\n\x05SHADE\x10\x04\x12\n\n\x06\x42UCKLE\x10\x05\x12\x10\n\x0c\x42LACK_CIRCLE\x10\x06\x12\x0e\n\nSHADE_BASE\x10\x07\x12\x08\n\x04\x42ULB\x10\x08\x12\x0c\n\x08\x42ULB_TOP\x10\t\"\xa6\x01\n\x08Sandwich\x12&\n\x05state\x18\x01 \x01(\x0e\x32\x17.gabriel.Sandwich.State\"r\n\x05State\x12\t\n\x05START\x10\x00\x12\x0b\n\x07NOTHING\x10\x01\x12\t\n\x05\x42READ\x10\x02\x12\x07\n\x03HAM\x10\x03\x12\x0b\n\x07LETTUCE\x10\x04\x12\x0c\n\x08\x43UCUMBER\x10\x05\x12\x08\n\x04HALF\x10\x06\x12\n\n\x06TOMATO\x10\x07\x12\x0c\n\x08HAMWRONG\x10\x08\"{\n\x0c\x45ngineFields\x12\x1c\n\x14last_update_frame_id\x18\x01 \x01(\x03\x12\x1d\n\x04ikea\x18\x02 \x01(\x0b\x32\r.gabriel.IkeaH\x00\x12%\n\x08sandwich\x18\x03 \x01(\x0b\x32\x11.gabriel.SandwichH\x00\x42\x07\n\x05stateB%\n\x1b\x65\x64u.cmu.cs.gabriel.protocolB\x06Protosb\x06proto3')
 )
 _sym_db.RegisterFileDescriptor(DESCRIPTOR)
 
@@ -74,8 +74,8 @@ _IKEA_STATE = _descriptor.EnumDescriptor(
   ],
   containing_type=None,
   options=None,
-  serialized_start=134,
-  serialized_end=266,
+  serialized_start=141,
+  serialized_end=273,
 )
 _sym_db.RegisterEnumDescriptor(_IKEA_STATE)
 
@@ -124,8 +124,8 @@ _SANDWICH_STATE = _descriptor.EnumDescriptor(
   ],
   containing_type=None,
   options=None,
-  serialized_start=321,
-  serialized_end=435,
+  serialized_start=328,
+  serialized_end=442,
 )
 _sym_db.RegisterEnumDescriptor(_SANDWICH_STATE)
 
@@ -171,8 +171,8 @@ _IKEA = _descriptor.Descriptor(
   extension_ranges=[],
   oneofs=[
   ],
-  serialized_start=24,
-  serialized_end=266,
+  serialized_start=31,
+  serialized_end=273,
 )
 
 
@@ -203,8 +203,8 @@ _SANDWICH = _descriptor.Descriptor(
   extension_ranges=[],
   oneofs=[
   ],
-  serialized_start=269,
-  serialized_end=435,
+  serialized_start=276,
+  serialized_end=442,
 )
 
 
@@ -251,8 +251,8 @@ _ENGINEFIELDS = _descriptor.Descriptor(
       name='state', full_name='gabriel.EngineFields.state',
       index=0, containing_type=None, fields=[]),
   ],
-  serialized_start=437,
-  serialized_end=560,
+  serialized_start=444,
+  serialized_end=567,
 )
 
 _IKEA.fields_by_name['state'].enum_type = _IKEA_STATE
@@ -273,21 +273,21 @@ DESCRIPTOR.message_types_by_name['EngineFields'] = _ENGINEFIELDS
 
 Ikea = _reflection.GeneratedProtocolMessageType('Ikea', (_message.Message,), dict(
   DESCRIPTOR = _IKEA,
-  __module__ = 'task_pb2'
+  __module__ = 'instruction_pb2'
   # @@protoc_insertion_point(class_scope:gabriel.Ikea)
   ))
 _sym_db.RegisterMessage(Ikea)
 
 Sandwich = _reflection.GeneratedProtocolMessageType('Sandwich', (_message.Message,), dict(
   DESCRIPTOR = _SANDWICH,
-  __module__ = 'task_pb2'
+  __module__ = 'instruction_pb2'
   # @@protoc_insertion_point(class_scope:gabriel.Sandwich)
   ))
 _sym_db.RegisterMessage(Sandwich)
 
 EngineFields = _reflection.GeneratedProtocolMessageType('EngineFields', (_message.Message,), dict(
   DESCRIPTOR = _ENGINEFIELDS,
-  __module__ = 'task_pb2'
+  __module__ = 'instruction_pb2'
   # @@protoc_insertion_point(class_scope:gabriel.EngineFields)
   ))
 _sym_db.RegisterMessage(EngineFields)
